requirements.txt
spidev~=3.4
requests~=2.21.0
Pillow~=5.4.1